refactor(mobile): useGTW hook and registration should use same code
