+++ apps/mobile/src/services/notifications/registration.ts
@@ -1,38 +1,21 @@
-import DeviceInfo from 'react-native-device-info'
-import { SiweMessage } from 'siwe'
 import { Wallet, HDNodeWallet } from 'ethers'
 
 import { store } from '@/src/store'
 import { selectSafeInfo } from '@/src/store/safesSlice'
 import { selectSigners } from '@/src/store/signersSlice'
 import { selectFirstDelegateForAnySafeOwner } from '@/src/store/delegatesSlice'
-import { NOTIFICATION_ACCOUNT_TYPE, ERROR_MSG } from '@/src/store/constants'
-import {
-  notificationChannels,
-  withTimeout,
-  getSigner,
-  REGULAR_NOTIFICATIONS,
-  OWNER_NOTIFICATIONS,
-} from '@/src/utils/notifications'
+import { NOTIFICATION_ACCOUNT_TYPE } from '@/src/store/constants'
+import { notificationChannels, withTimeout, getSigner } from '@/src/utils/notifications'
 import FCMService from './FCMService'
 import NotificationService from './NotificationService'
 import { setSafeSubscriptionStatus } from '@/src/store/safeSubscriptionsSlice'
-import { cgwApi as authApi } from '@safe-global/store/gateway/AUTO_GENERATED/auth'
-import { cgwApi as notificationsApi } from '@safe-global/store/gateway/AUTO_GENERATED/notifications'
 import Logger from '@/src/utils/logger'
-import { convertToUuid } from '@/src/utils/uuid'
-import { isAndroid } from '@/src/config/constants'
 import { getPrivateKey } from '@/src/hooks/useSign/useSign'
+import { registerForNotificationsOnBackEnd, unregisterForNotificationsOnBackEnd } from './backend'
 
 type DelegateInfo = { owner: string; delegateAddress: string } | null
 
 // Cache to track authenticated signers and their authentication timestamps
-type AuthCacheEntry = {
-  timestamp: number
-  chainId: string
-}
-const authCache: Record<string, AuthCacheEntry> = {}
-const AUTH_CACHE_EXPIRY_MS = 60000 // 60 seconds
 
 export const getDelegateKeyId = (safe: string, delegateAddress: string): string => `delegate_${safe}_${delegateAddress}`
 
@@ -45,146 +28,6 @@ export const getDelegateSigner = async (delegate: DelegateInfo) => {
   const privateKey = await getPrivateKey(delegateKeyId, { requireAuthentication: false })
   const signer = privateKey ? getSigner(privateKey) : null
   return { signer }
-}
-
-const getDeviceUuid = async () => {
-  const deviceId = await DeviceInfo.getUniqueId()
-  return convertToUuid(deviceId)
-}
-
-const getNotificationRegisterPayload = async ({
-  signer,
-  chainId,
-}: {
-  signer: Wallet | HDNodeWallet
-  chainId: string
-}) => {
-  // Add a cache buster to force a fresh request
-  const { nonce } = await store
-    .dispatch(
-      authApi.endpoints.authGetNonceV1.initiate(undefined, {
-        forceRefetch: true,
-      }),
-    )
-    .unwrap()
-
-  if (!nonce) {
-    throw new Error(ERROR_MSG)
-  }
-
-  const message = new SiweMessage({
-    address: signer.address,
-    chainId: Number(chainId),
-    domain: 'global.safe.mobileapp',
-    statement: 'Safe Wallet wants you to sign in with your Ethereum account',
-    nonce,
-    uri: 'https://safe.global',
-    version: '1',
-    issuedAt: new Date().toISOString(),
-  })
-
-  return { siweMessage: message.prepareMessage() }
-}
-
-const authenticateSigner = async (signer: Wallet | HDNodeWallet | null, chainId: string) => {
-  if (!signer) {
-    return
-  }
-
-  const signerAddress = signer.address
-  const cacheKey = `${signerAddress.toLowerCase()}`
-  const cachedAuth = authCache[cacheKey]
-
-  const now = Date.now()
-  if (cachedAuth && cachedAuth.chainId === chainId && now - cachedAuth.timestamp < AUTH_CACHE_EXPIRY_MS) {
-    // Use cached authentication if it's recent enough
-    Logger.info('Using cached authentication for signer', { signerAddress })
-    return
-  }
-
-  // Authenticate the signer
-  const { siweMessage } = await getNotificationRegisterPayload({ signer, chainId })
-  const signature = await signer.signMessage(siweMessage)
-
-  // Add cache buster to force a fresh request
-  await store
-    .dispatch(
-      authApi.endpoints.authVerifyV1.initiate({
-        siweDto: { message: siweMessage, signature },
-      }),
-    )
-    .unwrap()
-
-  // Update the cache with the new authentication timestamp
-  authCache[cacheKey] = {
-    timestamp: now,
-    chainId,
-  }
-  Logger.info('Authenticated signer and updated cache', { signerAddress })
-}
-
-export const registerForNotificationsOnBackEnd = async ({
-  safeAddress,
-  signer,
-  chainIds,
-  fcmToken,
-  notificationAccountType,
-}: {
-  safeAddress: string
-  signer: Wallet | HDNodeWallet | null
-  chainIds: string[]
-  fcmToken: string
-  notificationAccountType: NOTIFICATION_ACCOUNT_TYPE
-}) => {
-  const isOwner = notificationAccountType === NOTIFICATION_ACCOUNT_TYPE.OWNER
-  const deviceUuid = await getDeviceUuid()
-
-  await authenticateSigner(signer, chainIds[0])
-
-  const NOTIFICATIONS_GRANTED = isOwner ? OWNER_NOTIFICATIONS : REGULAR_NOTIFICATIONS
-
-  await store
-    .dispatch(
-      notificationsApi.endpoints.notificationsUpsertSubscriptionsV2.initiate({
-        upsertSubscriptionsDto: {
-          cloudMessagingToken: fcmToken,
-          safes: chainIds.map((chainId) => ({
-            chainId,
-            address: safeAddress,
-            notificationTypes: NOTIFICATIONS_GRANTED,
-          })),
-          deviceType: isAndroid ? 'ANDROID' : 'IOS',
-          deviceUuid,
-        },
-      }),
-    )
-    .unwrap()
-}
-
-export const unregisterForNotificationsOnBackEnd = async ({
-  signer,
-  safeAddress,
-  chainIds,
-}: {
-  signer: Wallet | HDNodeWallet | null
-  safeAddress: string
-  chainIds: string[]
-}) => {
-  await authenticateSigner(signer, chainIds[0])
-  const deviceUuid = await getDeviceUuid()
-
-  for (const chainId of chainIds) {
-    await store
-      .dispatch(
-        notificationsApi.endpoints.notificationsDeleteSubscriptionV2.initiate({
-          deviceUuid,
-          chainId,
-          safeAddress,
-        }),
-      )
-      .unwrap()
-    await withTimeout(Promise.resolve(), 200)
-  }
 }
 
 export const getNotificationAccountType = (safeAddress: string) => {
